Se vincula con la plataforma de GoogleMaps. Se muestra pantalla básica de Maps.

+++ app/src/main/java/com/ccastro/cooking/domain/models/Location.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Location(
     @SerializedName("urlFlagIcon") val urlImgBandera: String,
     @SerializedName("country") val pais: String,
-    @SerializedName("latitud") val lat: Float,
-    @SerializedName("longitud") val long: Float,
+    @SerializedName("latitud") val lat: Double,
+    @SerializedName("longitud") val long: Double,
     @SerializedName("region") val regionName: String,
 )
